addressed comments from PR review
