+++ h2o-automl/src/main/java/ai/h2o/automl/TrainingStepsProvider.java
@@ -10,7 +10,9 @@ public interface TrainingStepsProvider<T extends TrainingSteps> {
     String getName();
 
     /**
-     * @return the class providing the actual steps: this class must extend {@link TrainingSteps}.
+     *
+     * @param aml the {@link AutoML} instance needed to build the {@link TrainingSteps}
+     * @return an instance of {@link TrainingSteps} listing all the various AutoML steps executable with this provider name.
      */
-    Class<T> getStepsClass();
+    T newInstance(AutoML aml);
 }
