+++ h2o-automl/src/main/resources/META-INF/services/ai.h2o.automl.TrainingStepsProvider
@@ -1,6 +1,6 @@
-ai.h2o.automl.training.DeepLearningSteps$Provider
-ai.h2o.automl.training.DRFSteps$Provider
-ai.h2o.automl.training.GBMSteps$Provider
-ai.h2o.automl.training.GLMSteps$Provider
-ai.h2o.automl.training.StackedEnsembleSteps$Provider
-ai.h2o.automl.training.XGBoostSteps$Provider
+ai.h2o.automl.training.DeepLearningStepsProvider
+ai.h2o.automl.training.DRFStepsProvider
+ai.h2o.automl.training.GBMStepsProvider
+ai.h2o.automl.training.GLMStepsProvider
+ai.h2o.automl.training.StackedEnsembleStepsProvider
+ai.h2o.automl.training.XGBoostStepsProvider
